Remain compatible with older Linux versions
Newer Linux versions (Ubuntu 14.04, for one) have an updated GNU libc
that contains symbol references that are incompatible with older
versions. This commit pegs the references to the two symbols that
mmap-object uses to older versions.

The clock_gettime implementation was moved from librt to (GNU) libc with
GLIBC version 2.17. In order to resolve up the older symbol ref on newer
machines, librt is picked up explicitly.